# packages/api-contracts/openapi/workplace.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/skills/search": {
      "get": {
        "operationId": "ProfilesController_searchSkills",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/me/skills": {
      "put": {
        "operationId": "ProfilesController_replaceMySkills",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceProfileSkillsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/{userId}": {
      "get": {
        "operationId": "ProfilesController_getByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get profile by user id",
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/me": {
      "patch": {
        "operationId": "ProfilesController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update current user profile metadata",
        "tags": [
          "profiles"
        ]
      }
    },
    "/posts": {
      "post": {
        "description": "Creates a personal or group-scoped post.",
        "operationId": "PostsController_createPost",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created post payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/me": {
      "get": {
        "operationId": "PostsController_getMyPosts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated posts authored by current user"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my posts",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/user/{userId}": {
      "get": {
        "operationId": "PostsController_getPostsByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated posts authored by selected user"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List posts by user",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{postId}": {
      "patch": {
        "description": "Updates post content or visibility for the author.",
        "operationId": "PostsController_updatePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update post",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "description": "Deletes a post owned by the current user.",
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result status"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete post",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_getPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Canonical post payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get canonical post detail",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{postId}/share-link": {
      "get": {
        "description": "Returns web and app links for post sharing.",
        "operationId": "PostsController_getShareLink",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post share links"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get post share link",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{postId}/poll/vote": {
      "post": {
        "description": "Creates or updates current user poll vote on a post.",
        "operationId": "PostsController_votePoll",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VotePollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post payload including poll stats"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Vote in post poll",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/comments/posts/{postId}": {
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_listComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
...
    "/likes/posts/{postId}": {
      "post": {
        "operationId": "LikesController_likePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Likes"
        ]
      },
      "delete": {
        "operationId": "LikesController_unlikePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Likes"
        ]
      },
      "get": {
        "operationId": "LikesController_getLikeState",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Likes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Workplace API",
    "description": "Workplace backend API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReplaceProfileSkillsDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "bio": {
            "type": "string",
            "maxLength": 280
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "jobTitle": {
            "type": "string",
            "maxLength": 80
          },
          "organizationSize": {
            "type": "string",
            "maxLength": 40
          }
        }
      },
      "PostMediaInputDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid",
            "description": "A scanned private MediaObject uploaded by the post author."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Legacy media URL. New clients must send mediaId."
          },
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ]
          },
          "width": {
            "type": "number",
            "minimum": 1,
            "maximum": 12000
          },
          "height": {
            "type": "number",
            "minimum": 1,
            "maximum": 12000
          }
        },
        "required": [
          "type"
        ]
      },
      "CreatePostPollOptionDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "text"
        ]
      },
      "CreatePostPollDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 120
          },
          "options": {
            "minItems": 2,
            "maxItems": 6,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePostPollOptionDto"
            }
          }
        },
        "required": [
          "question",
          "options"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "followers",
              "private"
            ]
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "media": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMediaInputDto"
            }
          },
          "taggedUserIds": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "poll": {
            "$ref": "#/components/schemas/CreatePostPollDto"
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "followers",
              "private"
            ]
          }
        }
      },
      "VotePollDto": {
        "type": "object",
        "properties": {
          "optionId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "optionId"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}